Remove colorschemer plugin
There is now https://flathub.org/apps/me.hergert.Schemes for example, as
a standalone app.

If somebody is interested to revive the colorschemer plugin, the code
can be copied in its own repository.

For most gedit users, this plugin is not essential. It's better to
directly have really good style/color schemes shipped with
libgedit-gtksourceview (where there is currently room for improvement).

Removing the plugin also permits to reduce the amount of code that the
core developers need to maintain.

In libgedit-gtksourceview, the GtkSourceStyleSchemeManager API is being
reworked, so the colorschemer implementation needed to be adapted.